Revert "minor changes"
This reverts commit a3a92706435b1baa400ddfd6de6f2b585be6def7.
