+++ mobile/lib/screens/insights/insights_widgets.dart
@@ -7,6 +7,7 @@ import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/svg.dart';
+import 'dart:async';
 
 class InsightAirQualityWidget extends StatelessWidget {
   const InsightAirQualityWidget(this.insight, {super.key, required this.name});
@@ -501,13 +502,59 @@ class ForecastContainer extends StatelessWidget {
   }
 }
 
-class HealthTipsWidget extends StatelessWidget {
-  const HealthTipsWidget(this.insight, {super.key});
+class HealthTipsWidget extends StatefulWidget {
+  HealthTipsWidget(
+    this.insight,
+  );
   final Insight insight;
 
   @override
+  _HealthTipsWidgetState createState() => _HealthTipsWidgetState();
+}
+
+class _HealthTipsWidgetState extends State<HealthTipsWidget> {
+  late PageController _pageController;
+  late Timer _timer;
+  int _currentPage = 0;
+
+  @override
+  void initState() {
+    super.initState();
+    _pageController = PageController(initialPage: 0);
+    _startAutoScroll();
+  }
+
+  @override
+  void dispose() {
+    _pageController.dispose();
+    _stopAutoScroll();
+    super.dispose();
+  }
+
+  void _startAutoScroll() {
+    _timer = Timer.periodic(const Duration(seconds: 4), (_) {
+      if (_currentPage < widget.insight.healthTips.length - 1) {
+        _currentPage++;
+      } else {
+        _currentPage = 0;
+        _pageController.jumpToPage(
+            0); // Jump back to the first page for seamless scrolling
+      }
+      _pageController.animateToPage(
+        _currentPage,
+        duration: const Duration(milliseconds: 500),
+        curve: Curves.easeInOut,
+      );
+    });
+  }
+
+  void _stopAutoScroll() {
+    _timer.cancel();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    if (insight.healthTips.isEmpty) {
+    if (widget.insight.healthTips.isEmpty) {
       return AnimatedSwitcher(
         duration: const Duration(milliseconds: 500),
         child: Container(),
@@ -516,9 +563,7 @@ class HealthTipsWidget extends StatelessWidget {
 
     return AnimatedSwitcher(
       duration: const Duration(milliseconds: 500),
-      child: ListView(
-        physics: const NeverScrollableScrollPhysics(),
-        shrinkWrap: true,
+      child: Column(
         children: [
           const SizedBox(
             height: 32,
@@ -526,7 +571,7 @@ class HealthTipsWidget extends StatelessWidget {
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: 16),
             child: Text(
-              insight.healthTipsTitle(),
+              widget.insight.healthTipsTitle(),
               textAlign: TextAlign.left,
               style: CustomTextStyle.headline7(context),
             ),
@@ -536,20 +581,25 @@ class HealthTipsWidget extends StatelessWidget {
           ),
           SizedBox(
             height: 128,
-            child: ListView.builder(
-              scrollDirection: Axis.horizontal,
-              controller: ScrollController(),
+            child: PageView.builder(
+              controller: _pageController,
+              onPageChanged: (index) {
+                setState(() {
+                  _currentPage = index;
+                });
+              },
               itemBuilder: (context, index) {
                 return Padding(
                   padding: EdgeInsets.only(
                     left: index == 0 ? 12.0 : 6.0,
-                    right:
-                        index == (insight.healthTips.length - 1) ? 12.0 : 6.0,
+                    right: index == (widget.insight.healthTips.length - 1)
+                        ? 12.0
+                        : 6.0,
                   ),
-                  child: HealthTipContainer(insight.healthTips[index]),
+                  child: HealthTipContainer(widget.insight.healthTips[index]),
                 );
               },
-              itemCount: insight.healthTips.length,
+              itemCount: widget.insight.healthTips.length,
             ),
           ),
         ],
